[#1529] Move Chimera to OT::Mail Properties
Chimera is our custom class that extends Mustache to support both an
HTML file and equivalent TXT file. This is helpful for email templates
which ideally can provide content for both.
